SWATCH-599: Remove database integration from swatch-metrics
