+++ src/main/java/org/candlepin/subscriptions/metering/profile/OpenShiftWorkerProfile.java
@@ -31,7 +31,6 @@ import org.candlepin.subscriptions.metering.service.prometheus.PrometheusService
 import org.candlepin.subscriptions.metering.service.prometheus.config.PrometheusServiceConfiguration;
 import org.candlepin.subscriptions.metering.service.prometheus.promql.QueryBuilder;
 import org.candlepin.subscriptions.metering.task.MeteringTasksConfiguration;
-import org.candlepin.subscriptions.security.OptInController;
 import org.candlepin.subscriptions.task.TaskQueueProperties;
 import org.candlepin.subscriptions.task.queue.TaskConsumerConfiguration;
 import org.candlepin.subscriptions.task.queue.TaskProducerConfiguration;
@@ -69,10 +68,7 @@ import org.springframework.retry.support.RetryTemplateBuilder;
   MeteringTasksConfiguration.class
 })
 @ComponentScan(
-    basePackages = {
-      "org.candlepin.subscriptions.metering.api",
-      "org.candlepin.subscriptions.metering.retention"
-    },
+    basePackages = {"org.candlepin.subscriptions.metering.api"},
     // Prevent TestConfiguration annotated classes from being picked up by ComponentScan
     excludeFilters = {
       @ComponentScan.Filter(type = FilterType.CUSTOM, classes = TypeExcludeFilter.class),
@@ -158,7 +154,6 @@ public class OpenShiftWorkerProfile {
       QueryBuilder queryBuilder,
       PrometheusEventsProducer prometheusEventsProducer,
       @Qualifier("openshiftMetricRetryTemplate") RetryTemplate openshiftRetryTemplate,
-      OptInController optInController,
       @Qualifier("meteringBatchIdGenerator") SpanGenerator spanGenerator) {
     return new PrometheusMeteringController(
         clock,
@@ -167,7 +162,6 @@ public class OpenShiftWorkerProfile {
         queryBuilder,
         prometheusEventsProducer,
         openshiftRetryTemplate,
-        optInController,
         spanGenerator);
   }
 }
