+++ src/main/java/org/candlepin/subscriptions/tally/admin/InternalTallyResource.java
@@ -20,10 +20,14 @@
  */
 package org.candlepin.subscriptions.tally.admin;
 
+import io.micrometer.core.annotation.Timed;
+import jakarta.transaction.Transactional;
 import jakarta.ws.rs.BadRequestException;
 import java.time.OffsetDateTime;
+import java.time.temporal.ChronoUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.candlepin.subscriptions.ApplicationProperties;
+import org.candlepin.subscriptions.db.EventRecordRepository;
 import org.candlepin.subscriptions.db.model.config.OptInType;
 import org.candlepin.subscriptions.resource.ResourceUtils;
 import org.candlepin.subscriptions.retention.RemittanceRetentionController;
@@ -39,6 +43,7 @@ import org.candlepin.subscriptions.tally.admin.api.model.TallyResend;
 import org.candlepin.subscriptions.tally.admin.api.model.TallyResendData;
 import org.candlepin.subscriptions.tally.admin.api.model.TallyResponse;
 import org.candlepin.subscriptions.tally.admin.api.model.UuidList;
+import org.candlepin.subscriptions.tally.events.EventRecordsRetentionProperties;
 import org.candlepin.subscriptions.tally.job.CaptureSnapshotsTaskManager;
 import org.candlepin.subscriptions.util.ApplicationClock;
 import org.candlepin.subscriptions.util.DateRange;
@@ -52,6 +57,11 @@ import org.springframework.util.StringUtils;
 @Slf4j
 public class InternalTallyResource implements InternalApi {
 
+  public static final String FEATURE_NOT_ENABLED_MESSSAGE =
+      "This feature is not currently enabled.";
+  private static final String SUCCESS_STATUS = "Success";
+  private static final String REJECTED_STATUS = "Rejected";
+
   private final ApplicationClock clock;
   private final ApplicationProperties applicationProperties;
   private final MarketplaceResendTallyController resendTallyController;
@@ -62,10 +72,8 @@ public class InternalTallyResource implements InternalApi {
   private final InternalTallyDataController internalTallyDataController;
   private final SecurityProperties properties;
 
-  public static final String FEATURE_NOT_ENABLED_MESSSAGE =
-      "This feature is not currently enabled.";
-  private static final String SUCCESS_STATUS = "Success";
-  private static final String REJECTED_STATUS = "Rejected";
+  private final EventRecordRepository eventRecordRepository;
+  private final EventRecordsRetentionProperties eventRecordsRetentionProperties;
 
   @SuppressWarnings("java:S107")
   public InternalTallyResource(
@@ -77,7 +85,9 @@ public class InternalTallyResource implements InternalApi {
       TallyRetentionController tallyRetentionController,
       RemittanceRetentionController remittanceRetentionController,
       InternalTallyDataController internalTallyDataController,
-      SecurityProperties properties) {
+      SecurityProperties properties,
+      EventRecordRepository eventRecordRepository,
+      EventRecordsRetentionProperties eventRecordsRetentionProperties) {
     this.clock = clock;
     this.applicationProperties = applicationProperties;
     this.resendTallyController = resendTallyController;
@@ -87,6 +97,8 @@ public class InternalTallyResource implements InternalApi {
     this.remittanceRetentionController = remittanceRetentionController;
     this.internalTallyDataController = internalTallyDataController;
     this.properties = properties;
+    this.eventRecordRepository = eventRecordRepository;
+    this.eventRecordsRetentionProperties = eventRecordsRetentionProperties;
   }
 
   @Override
@@ -236,6 +248,20 @@ public class InternalTallyResource implements InternalApi {
     }
   }
 
+  @Override
+  @Transactional
+  @Timed("rhsm-subscriptions.events.purge")
+  public void purgeEventRecords() {
+    var eventRetentionDuration = eventRecordsRetentionProperties.getEventRetentionDuration();
+
+    OffsetDateTime cutoffDate =
+        OffsetDateTime.now().truncatedTo(ChronoUnit.DAYS).minus(eventRetentionDuration);
+
+    log.info("Purging event records older than {}", cutoffDate);
+    eventRecordRepository.deleteInBulkEventRecordsByTimestampBefore(cutoffDate);
+    log.info("Event record purge completed successfully");
+  }
+
   /** Update tally snapshots for all orgs */
   @Override
   public DefaultResponse tallyConfiguredOrgs() {
@@ -292,7 +318,6 @@ public class InternalTallyResource implements InternalApi {
   /**
    * Create or update an opt in configuration. This operation is idempotent
    *
-   * @param accountNumber
    * @param orgId
    * @return success or error message
    */
