+++ src/test/java/org/candlepin/subscriptions/metering/service/prometheus/PrometheusServiceTest.java
@@ -38,7 +38,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-@ActiveProfiles({"openshift-metering-worker", "test"})
+@ActiveProfiles("openshift-metering-worker")
 class PrometheusServiceTest implements ExtendWithPrometheusWiremock {
 
   @Autowired private PrometheusService service;
